Remove comment about the resource to protect
The resourceToProtect which corresponds to the resources attribute in
the OpenAM policy should be set according to your IG configuration.

If you are using a rebase in OpenIG, before the PEF, the
resourceToProtect must be the URL of the protected resource. Otherwise,
it should be the IG's URL to access to resource.